Skeleton build of libMOM6.a with a lot of shims
In order to use a verbatim MOM6 submodule, we need to stub out a lot
of FMS functionality using the "infra" shims. It would be nice to
replace MOM6's Fortran parameter file parser with a routine that lets
us supply a Python dictionary. For now, we need to at least hook up
the Python-Fortran interface using f2py.