Raise Fortran errors on Python side
The error handler shim learned a check_error() function, that will
test whether an error was raised by a call to MOM_error(FATAL,...).
In this case, we can figure out where the error occurred (with
relatively coarse granularity), and raise a Python RuntimeError, which
we can then use in the pytest framework, or general user code.