Return a dictionary of diagnostics for AG, if requested
The accelerate_ale and do_regrid calls learned how to interface with
AG's diagnostics. A list of diagnostics is available in
pyale.adapt_diags: any combination of these can be requested in a list
as the `diags` argument to accelerate_ale or do_regrid. In addition to
the regular H, (T, and S), a dictionary keyed by the diagnostic name
is returned. For the moment, the diagnostic arrays are full-size,
i.e. they include the external halos. This probably isn't desirable,
but it's a little bit easier than figuring out the grid extents.

Closes #1.