Allow resuming of accelerated ALE
The resume_ale function is provided, which takes the state from a
previous run of accelerate_ale, and restarts from there. Any
diagnostics present in the diagnostic dictionary are also
re-populated: they may be deleted, but additional diagnostics can't be
added.

Because we provide the output only on the computational domain, this
is the data used for the restart. Notably, halos are re-initialised to
zero. Hopefully this doesn't cause any issues.

Closes #3.