Respect default values of boolean params in file parser
Most types are interoperable between C and Fortran, so we can use a
direct reference to the Fortran parameter when we attempt to fill in a
value from a Python dictionary. However, booleans are very different,
so we have a different parameter that gets copied in. We need to make
sure this starts off with the default value, if it was supplied.

Closes #4.