Accelerate regridding/remapping with OpenMP
The remapping will be accelerated regardless of coordinate, but
there's only specific support for AG at the moment, unless the other
coordinates have their own OpenMP directives.

We're also pointing to a branch of AG which slightly modifies the
grid building routine to allocate some work arrays on the heap, to
work around a segfault which seems to be a stack overrun due to the
work arrays being duplicated for every thread. This is what we want,
but the segfault isn't!